pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalshiedge"
version = "0.1.0"
description = "Autonomous Kalshi prediction market trading agent"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pygooglenews>=0.1.0",
    "gnews>=0.3.0",
    "aiosqlite>=0.20.0",
    "python-telegram-bot>=21.0",
    "structlog>=24.0",
    "cryptography>=43.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"src/kalshiedge/dashboard.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
